pyproject.toml
[project]
name = "Eco_Friendly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.12",
    "langchain-astradb>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.2",
    "pypdf>=6.9.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.41.0",
]
